Fix losing postponed nodes due to node merging
Removing nodes was an issue, because when two stacks process the same
value only one will get postponed nodes.

The fix changes strategy of lifting postponed nodes to do not remove
nodes until parsing is finished.  Every node tracks number of children
(always in front of list of children) to be removed.