Harmonize trees with and without postponed nodes
Such trees should end up being different only by number of
postponed nodes.  In order to achieve this do contraction on
finishing and be more careful on moving postponed nodes.

This solves the issue and makes test pass, but we still might be
doing more work than needed which results in these inconsistencies
in presence of postponed nodes.