Show only one version for pure additions/deletions
The other version of the code doesn't add anything useful in this case
and omitting it makes output easier to comprehend.

The disadvantage is that hidden version might contain different
formatting, but that's more or less negligible.

There are a lot of changes and they aren't trivial, but it was hard to
make them, let alone commit separately.  The nice thing however is that
size of Print::print() method was almost cut in halve.