Less annoying cursor synchronization in gdiff
When cursor in one of views was on the only updated element in
both views, it used to be impossible to move it off that token on both
sides, because it was moved back automatically.

Now cursor is moved automatically only on switching views.  And in other
situations matching element is highlighted with additional line
highlighting which is independent of cursor position.